memory mapped data file

+++ storage/src/main/kotlin/uk/tvidal/kraft/storage/KRaftFileStorageConfig.kt
@@ -5,5 +5,5 @@ import java.nio.file.Path
 data class KRaftFileStorageConfig(
     val path: Path,
     val fileName: String = "kraft",
-    val fileSize: Int = DEFAULT_FILE_SIZE
+    val fileSize: Long = DEFAULT_FILE_SIZE
 )
